build working without cgo
