+++ components/register/all_cgo.go
@@ -1,0 +1,10 @@
+//go:build cgo
+package register
+
+import (
+	_ "go.viam.com/rdk/components/camera/register"
+	_ "go.viam.com/rdk/components/arm/register"
+	_ "go.viam.com/rdk/components/audioinput/register"
+	_ "go.viam.com/rdk/components/base/register"
+	_ "go.viam.com/rdk/components/gripper/register"
+)
