+++ robot/web/web.go
@@ -1,4 +1,3 @@
-//go:build cgo
 // Package web provides gRPC/REST/GUI APIs to control and monitor a robot.
 package web
 
@@ -8,7 +7,6 @@ import (
 	"html/template"
 	"io"
 	"io/fs"
-	"math"
 	"net"
 	"net/http"
 	"net/http/pprof"
@@ -28,8 +26,6 @@ import (
 	"github.com/jhump/protoreflect/dynamic"
 	"github.com/pkg/errors"
 	"github.com/rs/cors"
-	"github.com/viamrobotics/gostream"
-	streampb "github.com/viamrobotics/gostream/proto/stream/v1"
 	"go.opencensus.io/trace"
 	pb "go.viam.com/api/robot/v1"
 	"go.viam.com/utils"
@@ -40,8 +36,6 @@ import (
 	"goji.io/pat"
 	googlegrpc "google.golang.org/grpc"
 
-	"go.viam.com/rdk/components/audioinput"
-	"go.viam.com/rdk/components/camera"
 	"go.viam.com/rdk/config"
 	"go.viam.com/rdk/grpc"
 	"go.viam.com/rdk/module"
@@ -49,7 +43,6 @@ import (
 	"go.viam.com/rdk/robot"
 	grpcserver "go.viam.com/rdk/robot/server"
 	weboptions "go.viam.com/rdk/robot/web/options"
-	webstream "go.viam.com/rdk/robot/web/stream"
 	rutils "go.viam.com/rdk/utils"
 	"go.viam.com/rdk/web"
 )
@@ -187,40 +180,6 @@ func validSDPTrackName(name string) string {
 	return strings.ReplaceAll(name, ":", "+")
 }
 
-// refreshVideoSources checks and initializes every possible video source that could be viewed from the robot.
-func (svc *webService) refreshVideoSources() {
-	for _, name := range camera.NamesFromRobot(svc.r) {
-		cam, err := camera.FromRobot(svc.r, name)
-		if err != nil {
-			continue
-		}
-		existing, ok := svc.videoSources[validSDPTrackName(name)]
-		if ok {
-			existing.Swap(cam)
-			continue
-		}
-		newSwapper := gostream.NewHotSwappableVideoSource(cam)
-		svc.videoSources[validSDPTrackName(name)] = newSwapper
-	}
-}
-
-// refreshAudioSources checks and initializes every possible audio source that could be viewed from the robot.
-func (svc *webService) refreshAudioSources() {
-	for _, name := range audioinput.NamesFromRobot(svc.r) {
-		input, err := audioinput.FromRobot(svc.r, name)
-		if err != nil {
-			continue
-		}
-		existing, ok := svc.audioSources[validSDPTrackName(name)]
-		if ok {
-			existing.Swap(input)
-			continue
-		}
-		newSwapper := gostream.NewHotSwappableAudioSource(input)
-		svc.audioSources[validSDPTrackName(name)] = newSwapper
-	}
-}
-
 // A Service controls the web server for a robot.
 type Service interface {
 	resource.Resource
@@ -239,56 +198,6 @@ type Service interface {
 
 	// Returns the unix socket path the module server listens on.
 	ModuleAddress() string
-}
-
-// StreamServer manages streams and displays.
-type StreamServer struct {
-	// Server serves streams
-	Server gostream.StreamServer
-	// HasStreams is true if service has streams that require a WebRTC connection.
-	HasStreams bool
-}
-
-// New returns a new web service for the given robot.
-func New(r robot.Robot, logger golog.Logger, opts ...Option) Service {
-	var wOpts options
-	for _, opt := range opts {
-		opt.apply(&wOpts)
-	}
-	webSvc := &webService{
-		Named:        InternalServiceName.AsNamed(),
-		r:            r,
-		logger:       logger,
-		rpcServer:    nil,
-		streamServer: nil,
-		services:     map[resource.API]resource.APIResourceCollection[resource.Resource]{},
-		opts:         wOpts,
-		videoSources: map[string]gostream.HotSwappableVideoSource{},
-		audioSources: map[string]gostream.HotSwappableAudioSource{},
-	}
-	return webSvc
-}
-
-type webService struct {
-	resource.Named
-
-	mu                      sync.Mutex
-	r                       robot.Robot
-	rpcServer               rpc.Server
-	modServer               rpc.Server
-	streamServer            *StreamServer
-	services                map[resource.API]resource.APIResourceCollection[resource.Resource]
-	opts                    options
-	addr                    string
-	modAddr                 string
-	logger                  golog.Logger
-	cancelCtx               context.Context
-	cancelFunc              func()
-	isRunning               bool
-	activeBackgroundWorkers sync.WaitGroup
-
-	videoSources map[string]gostream.HotSwappableVideoSource
-	audioSources map[string]gostream.HotSwappableAudioSource
 }
 
 var internalWebServiceName = resource.NewName(
@@ -448,19 +357,6 @@ func (svc *webService) refreshResources() error {
 	return svc.updateResources(resources)
 }
 
-// Update updates the web service when the robot has changed.
-func (svc *webService) Reconfigure(ctx context.Context, deps resource.Dependencies, _ resource.Config) error {
-	svc.mu.Lock()
-	defer svc.mu.Unlock()
-	if err := svc.updateResources(deps); err != nil {
-		return err
-	}
-	if !svc.isRunning {
-		return nil
-	}
-	return svc.addNewStreams(svc.cancelCtx)
-}
-
 func (svc *webService) updateResources(resources map[resource.Name]resource.Resource) error {
 	// so group resources by API
 	groupedResources := make(map[resource.API]map[resource.Name]resource.Resource)
@@ -536,185 +432,6 @@ func (svc *webService) Close(ctx context.Context) error {
 	}
 	svc.activeBackgroundWorkers.Wait()
 	return err
-}
-
-func (svc *webService) streamInitialized() bool {
-	return svc.streamServer != nil && svc.streamServer.Server != nil
-}
-
-func (svc *webService) addNewStreams(ctx context.Context) error {
-	if !svc.streamInitialized() {
-		return nil
-	}
-	svc.refreshVideoSources()
-	svc.refreshAudioSources()
-	if svc.opts.streamConfig == nil {
-		if len(svc.videoSources) != 0 || len(svc.audioSources) != 0 {
-			svc.logger.Debug("not starting streams due to no stream config being set")
-		}
-		return nil
-	}
-
-	newStream := func(name string) (gostream.Stream, bool, error) {
-		// Configure new stream
-		config := *svc.opts.streamConfig
-		config.Name = name
-		stream, err := svc.streamServer.Server.NewStream(config)
-
-		// Skip if stream is already registered, otherwise raise any other errors
-		var registeredError *gostream.StreamAlreadyRegisteredError
-		if errors.As(err, &registeredError) {
-			return nil, true, nil
-		} else if err != nil {
-			return nil, false, err
-		}
-
-		if !svc.streamServer.HasStreams {
-			svc.streamServer.HasStreams = true
-		}
-		return stream, false, nil
-	}
-
-	for name, source := range svc.videoSources {
-		stream, alreadyRegistered, err := newStream(name)
-		if err != nil {
-			return err
-		} else if alreadyRegistered {
-			continue
-		}
-
-		svc.startVideoStream(ctx, source, stream)
-	}
-
-	for name, source := range svc.audioSources {
-		stream, alreadyRegistered, err := newStream(name)
-		if err != nil {
-			return err
-		} else if alreadyRegistered {
-			continue
-		}
-
-		svc.startAudioStream(ctx, source, stream)
-	}
-
-	return nil
-}
-
-func (svc *webService) makeStreamServer(ctx context.Context) (*StreamServer, error) {
-	svc.refreshVideoSources()
-	svc.refreshAudioSources()
-	var streams []gostream.Stream
-	var streamTypes []bool
-
-	if svc.opts.streamConfig == nil || (len(svc.videoSources) == 0 && len(svc.audioSources) == 0) {
-		if len(svc.videoSources) != 0 || len(svc.audioSources) != 0 {
-			svc.logger.Debug("not starting streams due to no stream config being set")
-		}
-		noopServer, err := gostream.NewStreamServer(streams...)
-		return &StreamServer{noopServer, false}, err
-	}
-
-	addStream := func(streams []gostream.Stream, name string, isVideo bool) ([]gostream.Stream, error) {
-		config := *svc.opts.streamConfig
-		config.Name = name
-		if isVideo {
-			config.AudioEncoderFactory = nil
-
-			// set TargetFrameRate to the framerate of the video source if available
-			props, err := svc.videoSources[name].MediaProperties(ctx)
-			if err != nil {
-				svc.logger.Warnw("failed to get video source properties", "name", name, "error", err)
-			} else if props.FrameRate > 0.0 {
-				// round float up to nearest int
-				config.TargetFrameRate = int(math.Ceil(float64(props.FrameRate)))
-			}
-			// default to 60fps if the video source doesn't have a framerate
-			if config.TargetFrameRate == 0 {
-				config.TargetFrameRate = 60
-			}
-
-			if runtime.GOOS == "windows" {
-				// TODO(RSDK-1771): support video on windows
-				svc.logger.Warnw("not starting video stream since not supported on Windows yet", "name", name)
-				return streams, nil
-			}
-		} else {
-			config.VideoEncoderFactory = nil
-		}
-		stream, err := gostream.NewStream(config)
-		if err != nil {
-			return streams, err
-		}
-		return append(streams, stream), nil
-	}
-	for name := range svc.videoSources {
-		var err error
-		streams, err = addStream(streams, name, true)
-		if err != nil {
-			return nil, err
-		}
-		streamTypes = append(streamTypes, true)
-	}
-	for name := range svc.audioSources {
-		var err error
-		streams, err = addStream(streams, name, false)
-		if err != nil {
-			return nil, err
-		}
-		streamTypes = append(streamTypes, false)
-	}
-
-	streamServer, err := gostream.NewStreamServer(streams...)
-	if err != nil {
-		return nil, err
-	}
-
-	for idx, stream := range streams {
-		if streamTypes[idx] {
-			svc.startVideoStream(ctx, svc.videoSources[stream.Name()], stream)
-		} else {
-			svc.startAudioStream(ctx, svc.audioSources[stream.Name()], stream)
-		}
-	}
-
-	return &StreamServer{streamServer, true}, nil
-}
-
-func (svc *webService) startStream(streamFunc func(opts *webstream.BackoffTuningOptions) error) {
-	waitCh := make(chan struct{})
-	svc.activeBackgroundWorkers.Add(1)
-	utils.PanicCapturingGo(func() {
-		defer svc.activeBackgroundWorkers.Done()
-		close(waitCh)
-		opts := &webstream.BackoffTuningOptions{
-			BaseSleep: 50 * time.Microsecond,
-			MaxSleep:  2 * time.Second,
-			Cooldown:  5 * time.Second,
-		}
-		if err := streamFunc(opts); err != nil {
-			if utils.FilterOutError(err, context.Canceled) != nil {
-				svc.logger.Errorw("error streaming", "error", err)
-			}
-		}
-	})
-	<-waitCh
-}
-
-func (svc *webService) startVideoStream(ctx context.Context, source gostream.VideoSource, stream gostream.Stream) {
-	// Honor ctx that may be coming from a Reconfigure.
-	ctxWithJPEGHint := gostream.WithMIMETypeHint(ctx, rutils.WithLazyMIMEType(rutils.MimeTypeJPEG))
-	svc.startStream(func(opts *webstream.BackoffTuningOptions) error {
-		streamVideoCtx, _ := utils.MergeContext(svc.cancelCtx, ctxWithJPEGHint)
-		return webstream.StreamVideoSource(streamVideoCtx, source, stream, opts, svc.logger)
-	})
-}
-
-func (svc *webService) startAudioStream(ctx context.Context, source gostream.AudioSource, stream gostream.Stream) {
-	svc.startStream(func(opts *webstream.BackoffTuningOptions) error {
-		// Merge ctx that may be coming from a Reconfigure.
-		streamAudioCtx, _ := utils.MergeContext(svc.cancelCtx, ctx)
-		return webstream.StreamAudioSource(streamAudioCtx, source, stream, opts, svc.logger)
-	})
 }
 
 // installWeb prepares the given mux to be able to serve the UI for the robot.
@@ -811,21 +528,8 @@ func (svc *webService) runWeb(ctx context.Context, options weboptions.Options) (
 		return err
 	}
 
-	svc.streamServer, err = svc.makeStreamServer(ctx)
-	if err != nil {
+	if err := svc.initStreamServer(ctx, options); err != nil {
 		return err
-	}
-	if err := svc.rpcServer.RegisterServiceServer(
-		ctx,
-		&streampb.StreamService_ServiceDesc,
-		svc.streamServer.Server.ServiceServer(),
-		streampb.RegisterStreamServiceHandlerFromEndpoint,
-	); err != nil {
-		return err
-	}
-	if svc.streamServer.HasStreams {
-		// force WebRTC template rendering
-		options.WebRTC = true
 	}
 
 	if options.Debug {
@@ -860,11 +564,7 @@ func (svc *webService) runWeb(ctx context.Context, options weboptions.Options) (
 				svc.logger.Errorw("error stopping rpc server", "error", err)
 			}
 		}()
-		if svc.streamServer.Server != nil {
-			if err := svc.streamServer.Server.Close(); err != nil {
-				svc.logger.Errorw("error closing stream server", "error", err)
-			}
-		}
+		svc.closeStreamServer()
 	})
 	svc.activeBackgroundWorkers.Add(1)
 	utils.PanicCapturingGo(func() {
