+++ services/register/all_cgo.go
@@ -1,0 +1,9 @@
+//go:build cgo
+package register
+
+import (
+	_ "go.viam.com/rdk/services/motion/register"
+	_ "go.viam.com/rdk/services/mlmodel/register"
+	_ "go.viam.com/rdk/services/navigation/register"
+	_ "go.viam.com/rdk/services/vision/register"
+)
