+++ web/server/entrypoint_notc.go
@@ -1,0 +1,11 @@
+//go:build !cgo
+
+package server
+
+import (
+	robotimpl "go.viam.com/rdk/robot/impl"
+)
+
+func createRobotOptions() []robotimpl.Option {
+	return []robotimpl.Option{}
+}
